refactor(core): 拆分了 admin,community (#246)
