--- noj-core/src/routes/admin.ts
@@ -1,637 +1,0 @@
-import { Hono } from "hono";
-import { adminMiddleware, authMiddleware } from "../middleware/auth.ts";
-import { parseJsonBody } from "../lib/request.ts";
-import { BadRequestError, ValidationError } from "../lib/errors.ts";
-import { listUsers } from "../services/auth.ts";
-import { getDashboardStats } from "../services/dashboard.ts";
-import { listAllProblems } from "../services/problems/problems.ts";
-import {
-  createJudgeImage,
-  deleteJudgeImage,
-  listJudgeImages,
-  updateJudgeImage,
-} from "../services/judge-images.ts";
-import type {
-  CreateJudgeImageInput,
-  UpdateJudgeImageInput,
-} from "../types/problems.ts";
-import {
-  deleteSubmission,
-  getSubmission,
-  listSubmissions,
-  rejudgeProblemSubmissions,
-  rejudgeSubmission,
-} from "../services/submissions/submissions.ts";
-import {
-  adminUpdateUserProfile,
-  banUser,
-  getUserBanHistory,
-  unbanUser,
-} from "../services/users.ts";
-import { addIpBan, listIpBans, removeIpBan } from "../services/banlist.ts";
-import {
-  listSettings,
-  resetSetting,
-  updateSetting,
-} from "../services/system-settings.ts";
-import { listAuditLogs } from "../services/audit-log.ts";
-import type { AuditAction } from "../types/audit-log.ts";
-import {
-  createRole,
-  deleteRole,
-  listPermissions,
-  listRoles,
-  updateRole,
-  updateUserRoles,
-} from "../services/admin-roles.ts";
-import { SUBMISSION_STATUSES } from "../types/index.ts";
-import {
-  addParticipants,
-  createContest,
-  deleteContest,
-  getContest,
-  getContestProblems,
-  listContests,
-  listParticipants,
-  removeParticipant,
-  updateContest,
-} from "../services/contest/contests.ts";
-import type {
-  CreateContestInput,
-  UpdateContestInput,
-} from "../types/contests.ts";
-import { isValidContestType } from "../types/contests.ts";
-import { buildPaginationMeta, parsePagination } from "../lib/pagination.ts";
-
-const router = new Hono<{ Variables: { userId: string; userRole: string } }>();
-
-// 路由组级中间件：所有 admin 端点均需认证 + 管理员权限。
-// 例外：公告管理端点（/announcements*）已抽至独立 router
-// （routes/admin-announcements.ts，细粒度权限 announcement:manage，
-// admin:full_access 通配放行或显式拥有该权限均可），此处对公告路径跳过
-// adminMiddleware——否则组级通配 use 会先行拦截细粒度权限持有者。
-router.use("*", authMiddleware, async (c, next) => {
-  if (c.req.path.startsWith("/api/v1/admin/announcements")) {
-    return next();
-  }
-  return await adminMiddleware(c, next);
-});
-
-// ─── 用户管理 ───────────────────────────────────────────────
-
-/**
- * 管理员获取用户列表（分页 + 搜索筛选）。
- * GET /api/v1/admin/users
- */
-router.get("/users", async (c) => {
-  let page = parseInt(c.req.query("page") ?? "1", 10);
-  let perPage = parseInt(c.req.query("per_page") ?? "20", 10);
-  if (isNaN(page) || page < 1) page = 1;
-  if (isNaN(perPage) || perPage < 1) perPage = 20;
-  if (perPage > 100) perPage = 100;
-
-  const keyword = c.req.query("keyword") || undefined;
-  // is_admin 筛选：true / false / 缺省（全部）
-  const isAdminParam = c.req.query("is_admin");
-  const isAdmin = isAdminParam === "true"
-    ? true
-    : isAdminParam === "false"
-    ? false
-    : undefined;
-  const from = c.req.query("from") || undefined;
-  const to = c.req.query("to") || undefined;
-
-  const result = await listUsers({ page, perPage, keyword, isAdmin, from, to });
-  return c.json({ data: result.data, pagination: result.pagination });
-});
-
-/**
- * 管理员修改用户角色分配。
- * PATCH /api/v1/admin/users/:id/role
- *
- * **BREAKING**: 旧格式 `{ "role": "admin"|"user" }` 不再接受。
- * 新格式: `{ "role_ids": ["<uuid>", ...] }`
- */
-router.patch("/users/:id/role", async (c) => {
-  const targetUserId = c.req.param("id") as string;
-  const body = await parseJsonBody<{ role_ids: string[] }>(c);
-
-  if (!body.role_ids || !Array.isArray(body.role_ids)) {
-    throw new ValidationError("缺少必填字段：role_ids（UUID 数组）");
-  }
-
-  await updateUserRoles(targetUserId, body.role_ids, c.get("userId"));
-
-  // 返回更新结果
-  return c.json({ data: { id: targetUserId, role_ids: body.role_ids } }, 200);
-});
-
-/**
- * 管理员编辑用户资料。
- * PUT /api/v1/admin/users/:id
- */
-router.put("/users/:id", async (c) => {
-  const targetUserId = c.req.param("id") as string;
-  const body = await parseJsonBody<{ email?: string; bio?: string }>(c);
-
-  if (body.email === undefined && body.bio === undefined) {
-    throw new BadRequestError("至少需要提供一个可更新字段（email 或 bio）");
-  }
-
-  const user = await adminUpdateUserProfile(targetUserId, body);
-  return c.json({ data: user }, 200);
-});
-
-// ─── 题目管理 ───────────────────────────────────────────────
-
-/**
- * 管理员获取全量题目列表（含 U 型和 P 型）。
- * GET /api/v1/admin/problems
- */
-router.get("/problems", async (c) => {
-  const page = parseInt(c.req.query("page") || "1", 10);
-  const limit = parseInt(c.req.query("limit") || "20", 10);
-
-  if (Number.isNaN(page) || Number.isNaN(limit)) {
-    throw new BadRequestError("分页参数 page 和 limit 必须为数字");
-  }
-
-  const result = await listAllProblems({
-    page: Math.max(1, page),
-    limit: Math.min(100, Math.max(1, limit)),
-    difficulty: c.req.query("difficulty") || undefined,
-    category_id: c.req.query("category_id") || undefined,
-    keyword: c.req.query("keyword") || undefined,
-  });
-
-  return c.json({
-    data: result.items,
-    total: result.total,
-    page: result.page,
-    limit: result.limit,
-  });
-});
-
-// ─── 提交管理 ───────────────────────────────────────────────
-
-/**
- * 管理员获取全部提交列表（分页 + 筛选）。
- * GET /api/v1/admin/submissions
- */
-router.get("/submissions", async (c) => {
-  let page = parseInt(c.req.query("page") ?? "", 10);
-  let perPage = parseInt(c.req.query("per_page") ?? "", 10);
-
-  if (isNaN(page) || page < 1) page = 1;
-  if (isNaN(perPage) || perPage < 1) perPage = 20;
-  if (perPage > 100) perPage = 100;
-
-  const userId = c.req.query("user_id") || undefined;
-  const userSearch = c.req.query("user_search") || undefined;
-  const problemId = c.req.query("problem_id") || undefined;
-  const problemSearch = c.req.query("problem_search") || undefined;
-  const submissionId = c.req.query("submission_id") || undefined;
-  const language = c.req.query("language") || undefined;
-  const status = c.req.query("status") || undefined;
-  const from = c.req.query("from") || undefined;
-  const to = c.req.query("to") || undefined;
-
-  const validStatuses = SUBMISSION_STATUSES;
-  if (status && !(validStatuses as readonly string[]).includes(status)) {
-    throw new BadRequestError(
-      `无效的状态值：${status}，有效值：${validStatuses.join("、")}`,
-    );
-  }
-
-  const result = await listSubmissions({
-    userId,
-    userSearch,
-    problemId,
-    problemSearch,
-    submissionId,
-    language,
-    status,
-    from,
-    to,
-    page,
-    perPage,
-  });
-
-  const totalPages = Math.ceil(result.total / perPage);
-
-  return c.json({
-    data: result.data,
-    pagination: {
-      page,
-      per_page: perPage,
-      total: result.total,
-      total_pages: totalPages,
-    },
-  });
-});
-
-/**
- * 管理员查看任意提交详情（含 code）。
- * GET /api/v1/admin/submissions/:id
- */
-router.get("/submissions/:id", async (c) => {
-  const id = c.req.param("id") as string;
-  // 传入 userId/userRole 确保管理员能看到 code/output/details
-  const result = await getSubmission(id, c.get("userId"), undefined, c);
-  return c.json({ data: result });
-});
-
-/**
- * 管理员删除提交记录。
- * DELETE /api/v1/admin/submissions/:id
- */
-router.delete("/submissions/:id", async (c) => {
-  const id = c.req.param("id") as string;
-  await deleteSubmission(id);
-  return c.body(null, 204);
-});
-
-/**
- * 管理员重测提交。
- * POST /api/v1/admin/submissions/:id/rejudge
- */
-router.post("/submissions/:id/rejudge", async (c) => {
-  const id = c.req.param("id") as string;
-  await rejudgeSubmission(id);
-  return c.json({ data: { message: "重测任务已提交", submission_id: id } });
-});
-
-/**
- * 管理员批量重测题目的所有已完结提交。
- * POST /api/v1/admin/problems/:id/rejudge
- */
-router.post("/problems/:id/rejudge", async (c) => {
-  const id = c.req.param("id") as string;
-  const result = await rejudgeProblemSubmissions(id);
-  return c.json({
-    data: {
-      message: "批量重测任务已提交",
-      problem_id: id,
-      ...result,
-    },
-  });
-});
-
-// ─── 竞赛管理 ───────────────────────────────────────────────
-
-router.get("/contests", async (c) => {
-  const { page, perPage } = parsePagination(c);
-  const typeQuery = c.req.query("type");
-  if (typeQuery && !isValidContestType(typeQuery)) {
-    throw new BadRequestError("竞赛类型不合法");
-  }
-  const type = typeQuery && isValidContestType(typeQuery)
-    ? typeQuery
-    : undefined;
-  const result = await listContests({ page, perPage, type, showAll: true });
-  return c.json({
-    data: result.data,
-    pagination: buildPaginationMeta(page, perPage, result.total),
-  });
-});
-
-router.get("/contests/:id", async (c) => {
-  const contestId = c.req.param("id") as string;
-  const [contest, problems] = await Promise.all([
-    getContest(contestId),
-    getContestProblems(contestId),
-  ]);
-  return c.json({ data: { ...contest, problems } });
-});
-
-router.post("/contests", async (c) => {
-  const body = await parseJsonBody<CreateContestInput>(c);
-  const data = await createContest(body, c.get("userId"));
-  return c.json({ data }, 201);
-});
-
-router.put("/contests/:id", async (c) => {
-  const body = await parseJsonBody<UpdateContestInput>(c);
-  const data = await updateContest(
-    c.req.param("id") as string,
-    body,
-  );
-  return c.json({ data });
-});
-
-router.delete("/contests/:id", async (c) => {
-  await deleteContest(c.req.param("id") as string);
-  return c.body(null, 204);
-});
-
-router.get("/contests/:id/participants", async (c) => {
-  const data = await listParticipants(c.req.param("id") as string);
-  return c.json({ data });
-});
-
-router.post("/contests/:id/participants", async (c) => {
-  const userIds = await parseJsonBody<string[]>(c);
-  if (!Array.isArray(userIds)) {
-    throw new BadRequestError("请求体必须为用户 ID 数组");
-  }
-  const added = await addParticipants(c.req.param("id") as string, userIds);
-  return c.json({ data: { added } }, 201);
-});
-
-router.delete("/contests/:id/participants/:userId", async (c) => {
-  await removeParticipant(
-    c.req.param("id") as string,
-    c.req.param("userId") as string,
-  );
-  return c.body(null, 204);
-});
-
-router.get("/contests/:id/submissions", async (c) => {
-  const contestId = c.req.param("id") as string;
-  await getContest(contestId);
-  const { page, perPage } = parsePagination(c);
-  const result = await listSubmissions({ contestId, page, perPage });
-  return c.json({
-    data: result.data,
-    pagination: buildPaginationMeta(page, perPage, result.total),
-  });
-});
-
-// ─── 评测镜像白名单管理 ───────────────────────────────────────
-
-/**
- * 获取所有白名单条目。
- * GET /api/v1/admin/judge-images
- */
-router.get("/judge-images", async (c) => {
-  const items = await listJudgeImages();
-  return c.json({ data: items });
-});
-
-/**
- * 新增白名单条目。
- * POST /api/v1/admin/judge-images
- */
-router.post("/judge-images", async (c) => {
-  const body = await parseJsonBody<CreateJudgeImageInput>(c);
-
-  if (!body.image?.trim()) {
-    throw new BadRequestError("镜像名不能为空");
-  }
-
-  const item = await createJudgeImage(body);
-  return c.json({ data: item }, 201);
-});
-
-/**
- * 更新白名单条目。
- * PUT /api/v1/admin/judge-images/:id
- */
-router.put("/judge-images/:id", async (c) => {
-  const id = c.req.param("id") as string;
-  const body = await parseJsonBody<UpdateJudgeImageInput>(c);
-  const item = await updateJudgeImage(id, body);
-  return c.json({ data: item }, 200);
-});
-
-/**
- * 删除白名单条目。
- * DELETE /api/v1/admin/judge-images/:id
- */
-router.delete("/judge-images/:id", async (c) => {
-  const id = c.req.param("id") as string;
-  await deleteJudgeImage(id);
-  return c.body(null, 204);
-});
-
-// ─── 仪表盘 ─────────────────────────────────────────────────
-
-/**
- * 仪表盘统计数据。
- * GET /api/v1/admin/dashboard/stats
- */
-router.get("/dashboard/stats", async (c) => {
-  const stats = await getDashboardStats();
-  return c.json({ data: stats });
-});
-
-// ─── 系统设置（issue #99）───────────────────────────────────
-
-/**
- * 列出所有系统设置项（DB-backed 5 项 + env-only N 项）。
- * GET /api/v1/admin/settings
- *
- * 注意：必须先注册静态路径 `/settings`，再注册参数化路径 `/settings/:key`，
- * 否则 `GET /settings` 会被 `/settings/:key` 误匹配。
- */
-router.get("/settings", async (c) => {
-  const items = await listSettings();
-  return c.json({ data: items });
-});
-
-/**
- * 更新单个设置项（UPSERT）。
- * PUT /api/v1/admin/settings/:key
- * body: { value: boolean | string }
- */
-router.put("/settings/:key", async (c) => {
-  const key = c.req.param("key") as string;
-  const body = await parseJsonBody<{ value: unknown }>(c);
-  if (!("value" in body)) {
-    throw new BadRequestError("请求体必须包含 value 字段");
-  }
-  const item = await updateSetting(key, body.value, c.get("userId"));
-  return c.json({ data: item }, 200);
-});
-
-/**
- * 重置单个设置项（DELETE FROM system_settings，回退到 env/default）。
- * DELETE /api/v1/admin/settings/:key
- * 幂等：DB 不存在也正常返回 204。
- */
-router.delete("/settings/:key", async (c) => {
-  const key = c.req.param("key") as string;
-  await resetSetting(key, c.get("userId"));
-  return c.body(null, 204);
-});
-
-// ─── IP 黑名单管理（issue #102）────────────────────────────────
-
-/**
- * 管理员列出 IP 黑名单（分页 + 模糊搜索）。
- * GET /api/v1/admin/blacklist?page=&per_page=&keyword=
- */
-router.get("/blacklist", async (c) => {
-  let page = parseInt(c.req.query("page") ?? "1", 10);
-  let perPage = parseInt(c.req.query("per_page") ?? "20", 10);
-  if (isNaN(page) || page < 1) page = 1;
-  if (isNaN(perPage) || perPage < 1) perPage = 20;
-  if (perPage > 100) perPage = 100;
-
-  const keyword = c.req.query("keyword") || undefined;
-  const result = await listIpBans({ page, perPage, keyword });
-  return c.json({ data: result.data, pagination: result.pagination });
-});
-
-/**
- * 管理员添加 IP 黑名单。
- * POST /api/v1/admin/blacklist
- * body: { ip_or_cidr, reason?, expires_at? }
- */
-router.post("/blacklist", async (c) => {
-  const body = await parseJsonBody<{
-    ip_or_cidr: string;
-    reason?: string;
-    expires_at?: string | null;
-  }>(c);
-  if (!body.ip_or_cidr) {
-    throw new ValidationError("缺少必填字段：ip_or_cidr");
-  }
-  const ban = await addIpBan(body, c.get("userId"));
-  return c.json({ data: ban }, 201);
-});
-
-/**
- * 管理员删除 IP 黑名单条目。
- * DELETE /api/v1/admin/blacklist/:id
- */
-router.delete("/blacklist/:id", async (c) => {
-  const id = c.req.param("id") as string;
-  await removeIpBan(id, c.get("userId"));
-  return c.body(null, 204);
-});
-
-// ─── 用户封禁管理（issue #102）────────────────────────────────
-
-/**
- * 管理员封禁用户。
- * PATCH /api/v1/admin/users/:id/ban
- * body: { reason?, banned_until? }
- */
-router.patch("/users/:id/ban", async (c) => {
-  const targetUserId = c.req.param("id") as string;
-  const body = await parseJsonBody<{
-    reason?: string;
-    banned_until?: string | null;
-  }>(c);
-  const user = await banUser(
-    targetUserId,
-    body.reason,
-    body.banned_until,
-    c.get("userId"),
-  );
-  return c.json({ data: user }, 200);
-});
-
-/**
- * 管理员解封用户。
- * PATCH /api/v1/admin/users/:id/unban
- */
-router.patch("/users/:id/unban", async (c) => {
-  const targetUserId = c.req.param("id") as string;
-  const user = await unbanUser(targetUserId, c.get("userId"));
-  return c.json({ data: user }, 200);
-});
-
-/**
- * 获取用户封禁历史（user-ban-table）。
- * GET /api/v1/admin/users/:id/bans
- */
-router.get("/users/:id/bans", async (c) => {
-  const targetUserId = c.req.param("id") as string;
-  const records = await getUserBanHistory(targetUserId);
-  return c.json({ data: records }, 200);
-});
-
-// ─── 审计日志（issue #101）───────────────────────────────────
-
-/**
- * 管理员查询审计日志（分页 + 筛选）。
- * GET /api/v1/admin/audit-logs
- *
- * Query:
- *   page, per_page (default 20, max 100)
- *   admin_id?, action?, from?, to?  (ISO 8601)
- */
-router.get("/audit-logs", async (c) => {
-  let page = parseInt(c.req.query("page") ?? "1", 10);
-  let perPage = parseInt(c.req.query("per_page") ?? "20", 10);
-  if (isNaN(page) || page < 1) page = 1;
-  if (isNaN(perPage) || perPage < 1) perPage = 20;
-  if (perPage > 100) perPage = 100;
-
-  const result = await listAuditLogs({
-    page,
-    perPage,
-    admin_id: c.req.query("admin_id") || undefined,
-    action: (c.req.query("action") || undefined) as AuditAction | undefined,
-    from: c.req.query("from") || undefined,
-    to: c.req.query("to") || undefined,
-  });
-  return c.json({ data: result.data, pagination: result.pagination });
-});
-
-// ─── 角色 RBAC 管理 ───────────────────────────────────────────
-
-/**
- * 管理员获取角色列表。
- * GET /api/v1/admin/roles
- */
-router.get("/roles", async (c) => {
-  const result = await listRoles();
-  return c.json({ data: result });
-});
-
-/**
- * 管理员创建角色。
- * POST /api/v1/admin/roles
- * body: { name, description?, parent_id?, permission_ids? }
- */
-router.post("/roles", async (c) => {
-  const body = await parseJsonBody<{
-    name: string;
-    description?: string;
-    parent_id?: string;
-    permission_ids?: string[];
-  }>(c);
-  const result = await createRole(body);
-  return c.json({ data: result }, 201);
-});
-
-/**
- * 管理员编辑角色。
- * PUT /api/v1/admin/roles/:id
- * body: { name?, description?, parent_id?, permission_ids? }
- */
-router.put("/roles/:id", async (c) => {
-  const id = c.req.param("id") as string;
-  const body = await parseJsonBody<{
-    name?: string;
-    description?: string;
-    parent_id?: string | null;
-    permission_ids?: string[];
-  }>(c);
-  const result = await updateRole(id, body);
-  return c.json({ data: result });
-});
-
-/**
- * 管理员删除角色。
- * DELETE /api/v1/admin/roles/:id
- */
-router.delete("/roles/:id", async (c) => {
-  const id = c.req.param("id") as string;
-  await deleteRole(id);
-  return c.body(null, 204);
-});
-
-/**
- * 管理员获取权限列表（按 resource 分组）。
- * GET /api/v1/admin/permissions
- */
-router.get("/permissions", async (c) => {
-  const result = await listPermissions();
-  return c.json({ data: result });
-});
-
-export default router;
